backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.9.0
pydantic-settings==2.5.2
sqlalchemy==2.0.23
aiosqlite==0.19.0
apscheduler==3.10.4
httpx==0.25.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
